# Serialize query JSON response in a stream.