--- a/pkg/util/marshal/marshal_test.go
+++ b/pkg/util/marshal/marshal_test.go
@@ -3,10 +3,14 @@ package marshal
 import (
 	"bytes"
 	"fmt"
+	"math/rand"
+	"reflect"
 	"testing"
+	"testing/quick"
 	"time"
 
 	json "github.com/json-iterator/go"
+	jsoniter "github.com/json-iterator/go"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/promql"
 	"github.com/prometheus/prometheus/promql/parser"
@@ -487,7 +491,7 @@ func Test_WriteQueryResponseJSON(t *testing.T) {
 		err := WriteQueryResponseJSON(logqlmodel.Result{Data: queryTest.actual}, &b)
 		require.NoError(t, err)
 
-		testJSONBytesEqual(t, []byte(queryTest.expected), b.Bytes(), "Query Test %d failed", i)
+		require.JSONEqf(t, queryTest.expected, b.String(), "Query Test %d failed", i)
 	}
 }
 
@@ -497,7 +501,7 @@ func Test_WriteLabelResponseJSON(t *testing.T) {
 		err := WriteLabelResponseJSON(labelTest.actual, &b)
 		require.NoError(t, err)
 
-		testJSONBytesEqual(t, []byte(labelTest.expected), b.Bytes(), "Label Test %d failed", i)
+		require.JSONEqf(t, labelTest.expected, b.String(), "Label Test %d failed", i)
 	}
 }
 
@@ -511,7 +515,7 @@ func Test_MarshalTailResponse(t *testing.T) {
 		bytes, err := json.Marshal(model)
 		require.NoError(t, err)
 
-		testJSONBytesEqual(t, []byte(tailTest.expected), bytes, "Tail Test %d failed", i)
+		require.JSONEqf(t, tailTest.expected, string(bytes), "Tail Test %d failed", i)
 	}
 }
 
@@ -567,7 +571,7 @@ func Test_LabelResponseMarshalLoop(t *testing.T) {
 		jsonOut, err := json.Marshal(r)
 		require.NoError(t, err)
 
-		testJSONBytesEqual(t, []byte(labelTest.expected), jsonOut, "Label Marshal Loop %d failed", i)
+		require.JSONEqf(t, labelTest.expected, string(jsonOut), "Label Marshal Loop %d failed", i)
 	}
 }
 
@@ -581,7 +585,7 @@ func Test_TailResponseMarshalLoop(t *testing.T) {
 		jsonOut, err := json.Marshal(r)
 		require.NoError(t, err)
 
-		testJSONBytesEqual(t, []byte(tailTest.expected), jsonOut, "Tail Marshal Loop %d failed", i)
+		require.JSONEqf(t, tailTest.expected, string(jsonOut), "Tail Marshal Loop %d failed", i)
 	}
 }
 
@@ -615,21 +619,122 @@ func Test_WriteSeriesResponseJSON(t *testing.T) {
 			err := WriteSeriesResponseJSON(tc.input, &b)
 			require.NoError(t, err)
 
-			testJSONBytesEqual(t, []byte(tc.expected), b.Bytes(), "Label Test %d failed", i)
+			require.JSONEqf(t, tc.expected, b.String(), "Label Test %d failed", i)
 		})
 	}
 }
 
-func testJSONBytesEqual(t *testing.T, expected []byte, actual []byte, msg string, args ...interface{}) {
-	var expectedValue map[string]interface{}
-	err := json.Unmarshal(expected, &expectedValue)
-	require.NoError(t, err)
+// wrappedValue and its Generate method is used by quick to generate a random
+// parser.Value.
+type wrappedValue struct {
+	parser.Value
+}
 
-	var actualValue map[string]interface{}
-	err = json.Unmarshal(actual, &actualValue)
-	require.NoError(t, err)
+func (w wrappedValue) Generate(rand *rand.Rand, size int) reflect.Value {
+	types := []string{
+		loghttp.ResultTypeMatrix,
+		loghttp.ResultTypeScalar,
+		loghttp.ResultTypeStream,
+		loghttp.ResultTypeVector,
+	}
+	t := types[rand.Intn(len(types))]
 
-	require.Equalf(t, expectedValue, actualValue, msg, args)
+	switch t {
+	case loghttp.ResultTypeMatrix:
+		s, _ := quick.Value(reflect.TypeOf(promql.Series{}), rand)
+		series, _ := s.Interface().(promql.Series)
+
+		l, _ := quick.Value(reflect.TypeOf(labels.Labels{}), rand)
+		series.Metric = l.Interface().(labels.Labels)
+
+		matrix := promql.Matrix{series}
+		return reflect.ValueOf(wrappedValue{matrix})
+	case loghttp.ResultTypeScalar:
+		q, _ := quick.Value(reflect.TypeOf(promql.Scalar{}), rand)
+		return reflect.ValueOf(wrappedValue{q.Interface().(parser.Value)})
+	case loghttp.ResultTypeStream:
+		stream := logproto.Stream{
+			Labels:  randLabels(rand).String(),
+			Entries: randEntries(rand),
+			Hash:    0,
+		}
+
+		streams := logqlmodel.Streams([]logproto.Stream{stream})
+		return reflect.ValueOf(wrappedValue{streams})
+	case loghttp.ResultTypeVector:
+		v, _ := quick.Value(reflect.TypeOf(promql.Sample{}), rand)
+		sample, _ := v.Interface().(promql.Sample)
+
+		l, _ := quick.Value(reflect.TypeOf(labels.Labels{}), rand)
+		sample.Metric = l.Interface().(labels.Labels)
+		vector := promql.Vector([]promql.Sample{sample})
+		return reflect.ValueOf(wrappedValue{vector})
+
+	}
+	return reflect.ValueOf(nil)
+}
+
+var letterRunes = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789_")
+
+func randLabel(rand *rand.Rand) labels.Label {
+	var label labels.Label
+	b := make([]rune, 10)
+	for i := range b {
+		b[i] = letterRunes[rand.Intn(len(letterRunes))]
+	}
+	b[0] = 'n'
+	label.Name = string(b)
+
+	for i := range b {
+		b[i] = letterRunes[rand.Intn(len(letterRunes))]
+	}
+	b[0] = 'v'
+	label.Value = string(b)
+
+	return label
+}
+
+func randLabels(rand *rand.Rand) labels.Labels {
+	var labels labels.Labels
+	for i := 0; i < rand.Intn(100); i++ {
+		labels = append(labels, randLabel(rand))
+	}
+
+	return labels
+}
+
+func randEntries(rand *rand.Rand) []logproto.Entry {
+	var entries []logproto.Entry
+	for i := 0; i < rand.Intn(100); i++ {
+		l, _ := quick.Value(reflect.TypeOf(""), rand)
+		entries = append(entries, logproto.Entry{Timestamp: time.Now(), Line: l.Interface().(string)})
+	}
+
+	return entries
+}
+
+func Test_EncodeResult_And_ResultValue_Parity(t *testing.T) {
+	f := func(w wrappedValue) bool {
+		var buf bytes.Buffer
+		js := jsoniter.NewStream(jsoniter.ConfigFastest, &buf, 0)
+		encodeResult(w.Value, js)
+		js.Flush()
+		actual := buf.String()
+
+		buf.Reset()
+		v, err := NewResultValue(w.Value)
+		require.NoError(t, err)
+		js.WriteVal(v)
+		js.Flush()
+		expected := buf.String()
+
+		require.JSONEq(t, expected, actual)
+		return true
+	}
+
+	if err := quick.Check(f, nil); err != nil {
+		t.Error(err)
+	}
 }
 
 func Benchmark_Encode(b *testing.B) {
@@ -638,6 +743,7 @@ func Benchmark_Encode(b *testing.B) {
 	for n := 0; n < b.N; n++ {
 		for _, queryTest := range queryTests {
 			require.NoError(b, WriteQueryResponseJSON(logqlmodel.Result{Data: queryTest.actual}, buf))
+			buf.Reset()
 		}
 	}
 }