**What this PR does / why we need it**:
Previously the query response was converted to `loghttp.ResultValue`. This requires a lot of allocations. Furthermore, the JSON response string was allocated before it was written to the response writer. This changes replaces the conversion from `parser.Value` to `loghttp.ResultValue` with a hand written JSON stream. This redueces the allocation overhead quite a bit:

```
› benchstat before.txt after.txt
name        old time/op    new time/op    delta
_Encode-16    51.5µs ± 1%    24.1µs ± 5%  -53.25%  (p=0.008 n=5+5)

name        old alloc/op   new alloc/op   delta
_Encode-16    7.68kB ± 0%    1.66kB ± 0%  -78.34%  (p=0.008 n=5+5)

name        old allocs/op  new allocs/op  delta
_Encode-16       129 ± 0%        28 ± 0%  -78.29%  (p=0.008 n=5+5)
```

The correctness is validate with quick check.

**Special notes for your reviewer**:

**Checklist**
- [ ] Reviewed the `CONTRIBUTING.md` guide
- [ ] Documentation added
- [x] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are documented in `docs/sources/upgrading/_index.md`